+++ Source/SampleEditView.cpp
@@ -215,8 +215,8 @@ void SampleEditView::createVolumeInputs()
     mVolumeSlider->setScrollWheelEnabled(false);
     mVolumeSlider->setMouseDragSensitivity(256);
     mVolumeSlider->setTextBoxStyle(Slider::TextBoxRight, false, 60, 32);
-    mVolumeSlider->valueFromTextFunction = [](const String& s) -> float { return Decibels::decibelsToGain(s.getFloatValue()); };
-    mVolumeSlider->textFromValueFunction = [](float v) -> String { return Decibels::toString(Decibels::gainToDecibels(v), 1); };
+    mVolumeSlider->valueFromTextFunction = [](const String& s) -> double { return Decibels::decibelsToGain(s.getFloatValue()); };
+    mVolumeSlider->textFromValueFunction = [](double v) -> String { return Decibels::toString(Decibels::gainToDecibels((float)v), 1); };
     mVolumeSlider->setValue(initialGain);
     mVolumeSlider->setChangeNotificationOnlyOnRelease(false);
     mVolumeSlider->onValueChange = [this]() {
